+++ R/learnitdown-package.R
@@ -1,12 +1,14 @@
 #' @details
 #' The SciViews learnitdown package provides additional R Markdown tags to build
-#' richer learning material with interactivity.
+#' richer learning material with interactivity and to manage H5P exercises,
+#' learnr tutorials and Shiny interactive applications with user identification
+#' and activity tracking in a MongoDB database.
 #'
 #' See [h5p()] or [launch_shiny()]
 #' @keywords internal
 "_PACKAGE"
 
-#' @importFrom learnr run_tutorial tutorial_options
+#' @importFrom learnr run_tutorial tutorial_options answer
 #' @importFrom remotes install_github
 #' @importFrom glue glue
 #' @importFrom webshot webshot
@@ -19,6 +21,9 @@
 #' @importFrom utils askYesNo compareVersion create.post packageVersion select.list URLdecode URLencode
 #' @importFrom jsonlite base64_dec base64_enc fromJSON toJSON
 #' @importFrom PKI PKI.digest PKI.decrypt PKI.encrypt
+#' @importFrom keyring key_set_with_value key_get
+#' @importFrom getPass getPass
+#' @importFrom stats runif
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start